Seperated classes to prevent circular imports
